backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.5.0
google-generativeai==0.7.2
googlemaps==4.10.0
pydantic==2.7.3
apscheduler==3.10.4
fpdf2==2.7.9
python-multipart==0.0.9
